Broke out pagination into a separate Svelte subcomponent
